DirectRecord and MultiRecordMap

+++ nfsdb-core/src/main/java/com/nfsdb/column/DirectInputStream.java
@@ -3,6 +3,6 @@ package com.nfsdb.column;
 import java.io.InputStream;
 
 public abstract class DirectInputStream extends InputStream {
-    public abstract int getLength();
+    public abstract long getLength();
     public abstract long copyTo(long address, long start, long length);
 }
